requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
yt-dlp>=2024.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scipy>=1.10.0
Pillow>=9.5.0
onnxruntime>=1.16.0
# Install CPU-only torch separately on Colab:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
boxmot>=10.0.0
